Tighten internal types: ScalarInt n_points, JAX-only Period/Age, kw-only
Continues the dtype-barrier work by promoting internal scalar metadata
to JAX-typed forms wherever it lives strictly inside pylcm:

- `UniformContinuousGrid.n_points` and `Piece.n_points` are stored as
  `jnp.int32` JAX scalars, converted from the Python literals at
  construction. `_init_uniform_grid` casts `start` / `stop` /
  `n_points` at the boundary before validation; the validator can then
  assume strict `ScalarFloat` / `ScalarInt` arguments and only check
  value invariants. Coordinate helpers (`linspace`, `logspace`,
  `get_*_coordinate`) tighten `n_points` to `ScalarInt` so the
  conversion happens once at the boundary instead of at every call.
- `Grid.get_coordinate` reverts to `ScalarFloat | Array` (no Python
  float). The single production caller in `regime_building/V.py`
  always passes a JAX array; tests that called the helpers with
  Python literals wrap them with `jnp.asarray` / `jnp.int32`.
- `Period` aliases `ScalarInt` and `Age` aliases `ScalarInt | ScalarFloat`
  for the JIT-internal scalar contexts. `AgeGrid.period_to_age` and
  `age_to_period` use plain `int | float` directly since they are
  user-facing API methods returning Python values.
- `_simulate_regime_in_period` and the `transitions.py` helpers now take
  `period: ScalarInt`. The simulation loop derives `period = jnp.int32
  (period_idx)` once per iteration and passes it through; dict-key
  lookups (`argmax_and_max_Q_over_a[period_idx]`,
  `period_to_regime_to_V_arr.get(period_idx + 1)`) keep using the
  Python int.
- `FlatRegimeParams` tightens to `MappingProxyType[str, Array]` —
  post-whitelist every leaf is a JAX array, the prior `bool | float |
  Array` union was stale.
- `safe_to_int32` renamed to `safe_to_int_dtype` to mirror
  `safe_to_float_dtype`.
- `_strip_V_arr_from_result` made fully kw-only.
- `pyproject.toml` ignores `ARG001` for
  `tests/test_float_dtype_invariants.py` so per-test
  `# noqa: ARG001` comments drop out and signatures collapse to a
  single line.
- `Piece` becomes `init=False` with a manual `__init__` that lifts
  `n_points` to `jnp.int32`, mirroring `UniformContinuousGrid`.

Test-side fallout addressed in the same commit: literals wrapped with
`jnp.asarray` / `jnp.int32` where helpers tightened, redundant
`# ty: ignore` comments dropped, and three "validator rejects
non-numeric" tests reframed to assert the boundary cast catches them.

+++ tests/test_grid_helpers.py
@@ -15,7 +15,9 @@ from tests.conftest import DECIMAL_PRECISION
 
 
 def test_linspace():
-    calculated = linspace(start=jnp.asarray(1.0), stop=jnp.asarray(2.0), n_points=6)
+    calculated = linspace(
+        start=jnp.asarray(1.0), stop=jnp.asarray(2.0), n_points=jnp.int32(6)
+    )
     expected = np.array([1, 1.2, 1.4, 1.6, 1.8, 2])
     aaae(calculated, expected, decimal=DECIMAL_PRECISION)
 
@@ -29,7 +31,7 @@ def test_linspace_mapped_value():
         value=jnp.asarray(1.2),
         start=start,
         stop=stop,
-        n_points=6,
+        n_points=jnp.int32(6),
     )
     assert np.allclose(calculated, 1.0)
 
@@ -41,7 +43,7 @@ def test_linspace_mapped_value():
         value=jnp.asarray(1.3),
         start=start,
         stop=stop,
-        n_points=6,
+        n_points=jnp.int32(6),
     )
     assert np.allclose(calculated, 1.5)
 
@@ -50,13 +52,15 @@ def test_linspace_mapped_value():
         value=jnp.asarray(0.6),
         start=start,
         stop=stop,
-        n_points=6,
+        n_points=jnp.int32(6),
     )
     assert np.allclose(calculated, -2.0)
 
 
 def test_logspace():
-    calculated = logspace(start=jnp.asarray(1.0), stop=jnp.asarray(100.0), n_points=7)
+    calculated = logspace(
+        start=jnp.asarray(1.0), stop=jnp.asarray(100.0), n_points=jnp.int32(7)
+    )
     expected = np.array(
         [
             1.0,
@@ -77,7 +81,7 @@ def test_logspace_mapped_value():
         value=jnp.asarray((2.15443469 + 4.64158883) / 2),
         start=jnp.asarray(1.0),
         stop=jnp.asarray(100.0),
-        n_points=7,
+        n_points=jnp.int32(7),
     )
     assert np.allclose(calculated, 1.5)
 
@@ -88,7 +92,7 @@ def test_map_coordinates_linear():
     grid_info = {
         "start": jnp.asarray(0.0),
         "stop": jnp.asarray(1.0),
-        "n_points": 3,
+        "n_points": jnp.int32(3),
     }
 
     grid = linspace(**grid_info)  # [0, 0.5, 1]
@@ -113,7 +117,7 @@ def test_map_coordinates_logarithmic():
     grid_info = {
         "start": jnp.asarray(1.0),
         "stop": jnp.asarray(2.0),
-        "n_points": 3,
+        "n_points": jnp.int32(3),
     }
 
     grid = logspace(**grid_info)  # [1.0, 1.414213562373095, 2.0]
@@ -138,7 +142,7 @@ def test_map_coordinates_linear_outside_grid():
     grid_info = {
         "start": jnp.asarray(1.0),
         "stop": jnp.asarray(2.0),
-        "n_points": 2,
+        "n_points": jnp.int32(2),
     }
 
     grid = linspace(**grid_info)  # [1, 2]
@@ -148,7 +152,7 @@ def test_map_coordinates_linear_outside_grid():
     # Get coordinates corresponding to values outside the grid [1, 2]
     coordinates = jnp.array(
         [
-            get_linspace_coordinate(value=jnp.asarray(grid_val), **grid_info)  # ty: ignore[no-matching-overload]
+            get_linspace_coordinate(value=jnp.asarray(grid_val), **grid_info)
             for grid_val in [-1.0, 0.0, 3.0]
         ]
     )
@@ -164,21 +168,23 @@ def test_get_linspace_coordinate_with_array():
         value=values,
         start=jnp.asarray(1.0),
         stop=jnp.asarray(2.0),
-        n_points=6,
+        n_points=jnp.int32(6),
     )
     expected = jnp.array([0.0, 1.0, 2.5])
     aaae(coords, expected, decimal=DECIMAL_PRECISION)
 
 
 def test_get_logspace_coordinate_with_array():
-    grid = logspace(start=jnp.asarray(1.0), stop=jnp.asarray(100.0), n_points=7)
+    grid = logspace(
+        start=jnp.asarray(1.0), stop=jnp.asarray(100.0), n_points=jnp.int32(7)
+    )
     mid = (float(grid[1]) + float(grid[2])) / 2
     values = jnp.array([mid])
     coords = get_logspace_coordinate(
         value=values,
         start=jnp.asarray(1.0),
         stop=jnp.asarray(100.0),
-        n_points=7,
+        n_points=jnp.int32(7),
     )
     aaae(coords, jnp.array([1.5]), decimal=DECIMAL_PRECISION)
